Split strings into separate variables
When running babel for Node 7+ compatibility, the test fails. This fixes it. I suspect it does some static analysis and cannot find the individual `messages` and `moduleNames` exports.

Thinking about it, it's because the default export is exposed in `exports.default` when we use babel, so it tries to get `exports.messages` instead of `exports.default.messages` (and the same for `exports.moduleNames`).