Implement backend connection logic and data replication manager
The Backend class connects to the Backend itself and exposes to it a simple API:
- info: get capture-lib version and the station's name
- replicate: triggers replication in the Replicator (accepts database names/urls, db username and password)

The Replicator class manages replication between a local and a remote database.
In our case, we will have two instances per station: one for the data DB, another for the log DB.
Call Replicator~replicate(dbName, dbUser, dbPass) and it will try to replicate
the local [something] database to the remote [something] db.
If the connection fails, reconnections are handled with an exponential backoff algorithm.

Both replicators and backend connector keep track of the connection status and publish it
in the backend under:
 - state - backend socket.io connection state
 - replicator:data:state - data db replicator state
 - replicator:log:state - log db replicator state