feat(parser): Use base64 encoding instead of quasi-binary
Why? From our testing it provides the same space savings and, worst case scenario, the packet is 1/3
larger than raw data (instead of being 2x the raw data like it was with quasy binary). All that and
quasi-binary did not play well with our Arduino and transceivers: it was very inconsistent about
data reception. There are no tests envolving this different encoding, nor how we hooked it up to the
parser, but it's standard (it's implemented inside NodeJS) and manual testing confirms it works.

BREAKING CHANGE: Packet encoding changed from quasi-binary to base64