fix(replicator): Prevent log replicator from going into an infinite loop
Scenario:
The log replicator has nothing else to send so it pauses itself.
All replicators log when they are active, paused, etc.
The log replicator logs that it has been paused.
Now it has something to send, so it becomes active.
It logs it's active.
Sends everything.
Becomes inactive again, and the cycle repeats.

It didn't seem to cause major damage but it would be very annoying to browse through
the logs if something bad happened. So we only log pause/active when an error occurs or
when we're entering them from a state that wasn't pause/active.